some new changes, todo per session added

+++ app.py
@@ -1,21 +1,31 @@
-from flask import Flask, render_template , request, redirect , jsonify
+from flask import Flask, jsonify, render_template , request, session
 from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
 from datetime import datetime
 import os
+import uuid
+import string
+import random
 
 
 
 app = Flask(__name__)
-db_path = os.path.join(os.path.dirname(__file__), 'app.db')
+db_path = os.path.join(os.path.dirname(__file__), 'sqlite3.db')
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///{}'.format(db_path)
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.secret_key = uuid.uuid4().hex
 
 db = SQLAlchemy(app)
+migrate = Migrate(app, db)
+
+def random_string(x=random.randrange(16, 32, 1)):
+    return ''.join(random.choice([*string.ascii_lowercase, *string.ascii_uppercase]) for _ in range(x))
 
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String(200), nullable=False)
     date_created = db.Column(db.DateTime,default=datetime.utcnow)
+    user_id = db.Column(db.String(200), nullable=False)
 
     def __repr__(self):
         return '<Task %r>' % self.id
@@ -26,44 +36,54 @@ class Todo(db.Model):
 def index():
     if request.method == 'POST':
         task_content = request.form['content']
-        new_task = Todo(content=task_content)
+        old_user = session.get('user_id')
+        user_id = old_user if bool(old_user) else random_string(32)
+        new_task = Todo(content=task_content, user_id=user_id)
         try:
             db.session.add(new_task)
             db.session.commit()
-            return redirect('/')
-        except:
-            return 'Error'
+            if not bool(old_user):
+                session['user_id'] = user_id
+            return jsonify({"success":"added", "content": request.form['content']}), 201
+        except Exception as e:
+            print(str(e))
+            return jsonify({"error":"couldn't added.", "content": request.form['content']}), 403
 
-    tasks = Todo.query.order_by(Todo.date_created).all()
+    tasks = Todo.query.filter_by(user_id=session.get('user_id')).order_by(Todo.date_created).all()
     return render_template('index.html',tasks=tasks)
 
-@app.route('/delete/<int:id>')
+@app.route('/delete/<int:id>', methods=['DELETE'])
 def delete(id):
-    task_to_delete = Todo.query.get_or_404(id)
+    task = Todo.query.filter_by(user_id=session.get('user_id')).all()
+    if bool(task):
+        task_to_delete = Todo.query.get_or_404(id)
 
-    try:
-        db.session.delete(task_to_delete)
-        db.session.commit()
-        return redirect('/')
-    except:
-        return 'An error occured'
+        try:
+            db.session.delete(task_to_delete)
+            db.session.commit()
+            return jsonify({"success":"deleted"}), 204
+        except:
+            return 'An error occured'
+    return jsonify({"error":"This task doesn't belong to you"}), 400
 
 
-@app.route('/update/<int:id>', methods=['GET','POST'])
+@app.route('/update/<int:id>', methods=['PUT'])
 def update(id):
     task = Todo.query.get_or_404(id)
 
-    if request.method == 'POST':
-        task.content = request.form['content']
+    if request.method == 'PUT':
+        task = Todo.query.filter_by(user_id=session.get('user_id')).all()[0]
+        if bool(task):
+            task.content = request.form['content']
 
-        try:
-            db.session.commit()
-            return redirect('/')
+            try:
+                db.session.commit()
+                return jsonify({"success":"updated", 'content': request.form['content']}), 201
 
-        except:
-            return 'An error occured'
+            except:
+                return 'An error occured'
 
-    return render_template('update.html',task=task)
+    return jsonify({"error":"This task doesn't belong to you"}), 400
